feat: rename helm command to helm-local
