+++ README.md
@@ -69,7 +69,7 @@ pytest
 ## 문서
 
 각 명령어에 대한 자세한 문서는 `docs/` 디렉토리에서 확인할 수 있습니다:
-- [Helm 명령어](docs/helm.md)
+- [Helm Local 명령어](docs/helm-local.md)
 - [Docker Tar 명령어](docs/docker_tar.md)
 - [FatPack 명령어](docs/fatpack.md)
 - [S3 공유 명령어](docs/s3-share.md)
